feat: add timeout to async tests
